requirements.txt
flask==3.0.0
flask-cors==4.0.0
supabase==2.3.0
pandas==2.1.4
matplotlib==3.8.2
reportlab==4.0.7
pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
